pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = "LingoSummar - Intelligent text summarization API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.1",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "numpy>=2.4.3",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.8.0",
    "spacy>=3.8.11",
    "sqlmodel>=0.0.37",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning:importlib._bootstrap",
]
